fix warmup ratio

+++ src/main.py
@@ -68,7 +68,7 @@ if __name__ == "__main__":
     parser.add_argument('--cuda', action='store_true')
     parser.add_argument('--num_devices', type=int, default=1)
     parser.add_argument('--lr', type=float, default=3e-5)
-    parser.add_argument('--warmup_ratio', type=int, default=0.01)
+    parser.add_argument('--warmup_ratio', type=float, default=0.01)
     parser.add_argument('--precision', type=str, default='32')
     parser.add_argument('--dataset_name_or_path', type=str, default='duongttr/chebi-20-new')
     parser.add_argument('--model_config', type=str, default='src/configs/config_use_v_nofg.yaml')
